Fix: Have Animated object use latest state
Write a workaround by adding a listener to the Animated.Value object to
be able to listen for the end of a gesture before invoking a callback
function. This was the only way I could find that animations could use
the most up-to-date state since they are mounted before JavaScript is
run.